Change semantics of `advance!` to avoid changing the cursor.
The original semantics (leaving the cursor pointing to the end
of the stream after raising an error) weren't that useful in practice.
It's much more useful to leave the cursor unchanged in the case
that `advance!` fails to have enough `bytes_ahead` to move forward
by the requested amount, and raises an `error!`.

This is a breaking change in the semantics of an existing method,
but existing code using this method is not relying on those semantics,
so in practice it should not break existing code in known libraries.